A hack by Till:
The motor record is a complicated state machine
that is very hard to understand (partly because it is
not designed like a proper state-machine).
It is IMHO *impossible* to ensure that there are no bugs.

State transitions can happen asynchronously and
due to different events such as database-puts,
record-processing or modifications by device-support/drivers.

The logic that controls state transitions is very complicated
and not cleanly designed but shows signs of evolutionary
spaghetti-code.

That said - I believe I found the race-condition which prevents
DMOV from being set during the boot sequence:

If the motor record finds that the motor is not moving
anymore (movn=0) it internally sets dmov, does a few things
and finally, if DLY is > 0, then it internally clears dmov
again and schedules a callback for the record to be processed
once more in the future (after DLY expires). This would eventually
set DMOV=1.

However, it is possible that the callback is effectively canceled
e.g., by setting the STOP field between the callback being
scheduled and it going off. (There may well be other scenarios
where the callback is canceled.) Since the callback never
happens DMOV is never set.

I added a hack so that under the specified conditions the
callback is not canceled but left pending. I could successfully
reboot b131-iocmotor with both gap motors ending up having DMOV==1.

Disabling the hack resulted in the bad state where the lower
gap motor ends up with DMOV==0. I could switch back and forth
a few times and consistently got working/not-working behavior
depending on the hack being enabled/disabled.