Now handles the stall mode where a motor is commanded to move/home, but has trouble to do so.  When this is detected, we will get an error code, the motor record will alarm.
Depends on the SM setting, the motor hardware will stop, or keep trying
hard -- it will draw the maximum current allowed and probably heat up.

The desired setting for SM is "Stop on stall".  When this is detected,
we need to find out why it has trouble to move, instead of blindly forcing
it to try, and maybe even increase the run current.


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/epics/trunk/modules/motor/current@10438 ececa60f-9f63-0410-97c7-ea05888eb260